+++ iModelCore/ECDb/ECDb/ECDbImpl.cpp
@@ -265,6 +265,7 @@ void ECDb::Impl::RegisterECSqlPragmas() const
     GetPragmaManager().Register(PragmaPurgeOrphanRelationships::Create());
     GetPragmaManager().Register(PragmaDbList::Create());
     GetPragmaManager().Register(PragmaCheckECSqlWriteValues::Create());
+    GetPragmaManager().Register(PragmaECSqlVersion::Create());
     }
 
 //--------------------------------------------------------------------------------------